Second level of serialization
Now, I can add update data. The problem was that the object wasn't
updated before the serialization, so we didn't get the attributes like
the name, or the number of the model when we serialize the objects.

Everything is done now! Let's improve the serialization method and make
it more secure and robust!