Unchanging the modification of the self.__object state
In the previous code, when I was saving the data, I was changing the intern
state of the self.__object with the new_object ready for serialisation. That causes
that in the code, for the next object saving, there's a problem when new new_object
variables are made to save new variables into the json file.

So I cleaned the wrong code to add make it more clear

@Valeryio